Checking out the Tab layout style
So this is the start of a new project for inventory management. At this
point, i am undecided on how the Lyout for the User interface would look
like so i decided to try the TabPane layout to see how it goes. The
layout would be decided on by the next commit.