Link up with Database going smoothly
I should have used the JDBC API right from the start. Hibernate was
meant for complex use cases but well in the interest of learning we had
to try shey?? lol anyway the integration with the database is almost
complete. i just have to replicate what i did now on the Sale Dialog and
existing Products Dialog.